+++ artemis-test/src/main/java/com/artemis/system/BasicVoidSystem.java
@@ -5,6 +5,7 @@ import com.artemis.component.TransPackedInt;
 import com.artemis.managers.TagManager;
 import com.artemis.systems.VoidEntitySystem;
 
+@SuppressWarnings("unused")
 public class BasicVoidSystem extends VoidEntitySystem {
 
 	private TagManager manager;