+++ artemis-test/src/main/java/com/artemis/system/WiredBasicVoidSystem.java
@@ -7,6 +7,7 @@ import com.artemis.managers.TagManager;
 import com.artemis.systems.VoidEntitySystem;
 
 @Wire
+@SuppressWarnings("unused")
 public class WiredBasicVoidSystem extends VoidEntitySystem {
 
 	private TagManager tagManager;